security: remove the security_netlink_recv hook as it is equivalent to capable()
Once upon a time netlink was not sync and we had to get the effective
capabilities from the skb that was being received.  Today we instead get
the capabilities from the current task.  This has rendered the entire
purpose of the hook moot as it is now functionally equivalent to the
capable() call.

Change-Id: Id19d37757d793af068cf2bb54f009e382561979d
Signed-off-by: Eric Paris <eparis@redhat.com>