tcp: add a sysctl to config the tcp_default_init_rwnd
The default initial rwnd was hardcoded to 10.

Now we allow it to be controlled via
  /proc/sys/net/ipv4/tcp_default_init_rwnd
which limits the values from 3 to 100

This is somewhat needed because ipv6 routes are
autoconfigured by the kernel.

See "An Argument for Increasing TCP's Initial Congestion Window"
in https://developers.google.com/speed/articles/tcp_initcwnd_paper.pdf
Signed-off-by: JP Abgrall <jpa@google.com>

Change-Id: I52b1caa562cc8989e172798b8bb9b6486809c51a